llm_server/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118
transformers==4.36.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
llama-cpp-python==0.2.20
python-multipart==0.0.6
requests==2.31.0
tqdm==4.66.1
ninja==1.11.1.2
